+++ src/main/scala/com/ovoenergy/datastore4s/FieldFormat.scala
@@ -6,10 +6,14 @@ trait FieldFormat[A] {
 
   def toEntityField(fieldName: String, value: A): Field
 
-  def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, A] =
-    fromEntityFieldWithoutContext(fieldName, entity).left.map(DatastoreError.errorInField(fieldName))
+  def fromEntityFieldWithContext(fieldName: String, entity: Entity): Either[DatastoreError, A] =
+    fromEntityField(fieldName, entity).left.map(DatastoreError.errorInField(fieldName))
 
-  private[datastore4s] def fromEntityFieldWithoutContext(fieldName: String, entity: Entity): Either[DatastoreError, A]
+  def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, A]
+
+  def toValue[B](scalaValue: B)(implicit valueFormat: ValueFormat[B]): DatastoreValue = valueFormat.toValue(scalaValue)
+
+  def ignoreIndexes: FieldFormat[A] = FieldFormat.ignoreIndexes(this)
 
 }
 
@@ -18,11 +22,12 @@ final case class Field private (values: Map[String, DatastoreValue]) {
 
   def +(other: Field) = Field(other.values ++ values) // Composite field
 
-  def ignoreIndexes = Field(values.map { case (key, value) => (key, value.ignoreIndexes) })
+  def ignoreIndexes = Field(values.map { case (key, value) => (key, value.ignoreIndex) })
 }
 
 object Field {
   def apply(name: String, value: DatastoreValue): Field = new Field(Map(name -> value))
+  def apply(fields: (String, DatastoreValue)*): Field = new Field(fields.toMap)
 }
 
 object FieldFormat {
@@ -32,7 +37,7 @@ object FieldFormat {
       override def toEntityField(fieldName: String, value: A) =
         Field(fieldName, valueFormat.toValue(value))
 
-      override def fromEntityFieldWithoutContext(fieldName: String, entity: Entity): Either[DatastoreError, A] =
+      override def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, A] =
         entity
           .field(fieldName)
           .map(valueFormat.fromValue)
@@ -48,14 +53,20 @@ object FieldFormat {
         case Right(r) => rightFormat.toEntityField(fieldName, r) + (s"$fieldName.$eitherField", StringValue("Right"))
       }
 
-      override def fromEntityFieldWithoutContext(fieldName: String, entity: Entity): Either[DatastoreError, Either[L, R]] =
+      override def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, Either[L, R]] =
         entity.field(s"$fieldName.$eitherField") match {
-          case Some(StringValue("Left"))  => leftFormat.fromEntityField(fieldName, entity).map(Left(_))
-          case Some(StringValue("Right")) => rightFormat.fromEntityField(fieldName, entity).map(Right(_))
+          case Some(StringValue("Left"))  => leftFormat.fromEntityFieldWithContext(fieldName, entity).map(Left(_))
+          case Some(StringValue("Right")) => rightFormat.fromEntityFieldWithContext(fieldName, entity).map(Right(_))
           case Some(other)                => DatastoreError.error(s"Either field should be either 'Left' or 'Right' but was $other.")
           case None                       => DatastoreError.missingField(eitherField, entity)
         }
     }
+
+  def ignoreIndexes[A](existingFormat: FieldFormat[A]): FieldFormat[A] = new FieldFormat[A] {
+    override def toEntityField(fieldName: String, value: A) = existingFormat.toEntityField(fieldName, value).ignoreIndexes
+
+    override def fromEntityField(fieldName: String, entity: Entity) = existingFormat.fromEntityField(fieldName, entity)
+  }
 
   import scala.language.experimental.macros
 
@@ -87,7 +98,7 @@ object FieldFormat {
       if (helper.isObject(subType)) {
         cq"""Right(${subType.name.toString}) => Right(${helper.singletonObject(subType)})"""
       } else {
-        cq"""Right(${subType.name.toString}) => FieldFormat[$subType].fromEntityField(fieldName, entity)"""
+        cq"""Right(${subType.name.toString}) => FieldFormat[$subType].fromEntityFieldWithContext(fieldName, entity)"""
       }
     }
 
@@ -99,7 +110,7 @@ object FieldFormat {
               case ..$toCases
             }
 
-            override def fromEntityFieldWithoutContext(fieldName: String, entity: Entity): Either[DatastoreError, $fieldType] = stringFormat.fromEntityField(fieldName + ".type", entity) match {
+            override def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, $fieldType] = stringFormat.fromEntityFieldWithContext(fieldName + ".type", entity) match {
               case ..$fromCases
               case Right(other) => DatastoreError.error(s"Unknown subtype found: $$other")
               case Left(error) => Left(error)
@@ -127,7 +138,7 @@ object FieldFormat {
 
     val fieldFormats = fields.map { field =>
       val fieldName = field.asTerm.name
-      fq"""${field.name} <- implicitly[FieldFormat[${field.typeSignature}]].fromEntityField(fieldName + "." + ${fieldName.toString}, entity)"""
+      fq"""${field.name} <- implicitly[FieldFormat[${field.typeSignature}]].fromEntityFieldWithContext(fieldName + "." + ${fieldName.toString}, entity)"""
     }
 
     context.Expr[FieldFormat[A]](q"""import com.ovoenergy.datastore4s._
@@ -137,7 +148,7 @@ object FieldFormat {
               ${fieldExpressions.reduce(concatFieldExpressionsWithAdd(context)(_, _))}
             }
 
-            override def fromEntityFieldWithoutContext(fieldName: String, entity: Entity): Either[DatastoreError, $fieldType] = {
+            override def fromEntityField(fieldName: String, entity: Entity): Either[DatastoreError, $fieldType] = {
               for (
                 ..$fieldFormats
               ) yield $companion.apply(..$companionNamedArguments)